Added buffer fuddling to thwart CLANG's optimizer
The C/C++ standards I'm sure say that reading from an alloc'd buffer is "undefined behavior" which the compiler can optimize out, but that is absurd, and should never have been the case, and it's absurd that compilers do it.  Anything that is well specified on the CPU I'm using should run, period.  OF COURSE YOU CAN READ FROM MEMORY YOU JUST ALLOCATED.  It has a value before you've written to it, folks, I hate to break it to you.  I wanted to see what it was.  What is wrong with you people???